moved the logic of the database posts, comments  to src/db
