+++ src/app/actions/posts.tsx
@@ -1,47 +1,21 @@
 'use server';
-import { db } from '@/db';
-import { z } from 'zod';
 
-const createPostSchema = z.object({
-  title: z.string().min(3, "Title should be longer than 3 letters "),
-  content: z.string().min(10, "Content should be longer than 10 letters"),
-  userId: z.string().min(1, "User ID is required"),
-  topicId: z.string().min(1, "Topic ID is required"),
-});
+import { createPostInDB, deletePostFromDB, getPopularPostsFromDB } from '@/db/posts';
 
 export const createPost = async (_prevState: { message: string }, formData: FormData) => {
-  const title = formData.get('title');
-  const content = formData.get('content');
-  const userId = formData.get('userId');
-  const topicId = formData.get('topicId');
-
   try {
-    const parsedData = createPostSchema.parse({ title, content, userId, topicId });
-
-    const createdPost = await db.post.create({
-      data: parsedData,
-    });
-
-    console.log('Created post:', createdPost);
-
-    return { message: 'Post created successfully' };
-  } catch (error: unknown) {
+    const result = await createPostInDB(formData);
+    console.log('Created post:', result.createdPost);
+    return { message: result.message };
+  } catch (error) {
     console.error('Error creating post:', error);
-    if (error instanceof z.ZodError) {
-      return { message: error.errors[0].message };
-    } else if (error instanceof Error) {
-      return { message: error.message };
-    } else {
-      return { message: 'Something went wrong' };
-    }
+    return { message: 'Something went wrong' };
   }
 };
 
 export const deletePost = async (postId: string) => {
   try {
-    await db.post.delete({
-      where: { id: postId },
-    });
+    await deletePostFromDB(postId);
   } catch (error) {
     console.error('Error deleting post:', error);
     throw new Error('Failed to delete post');
@@ -50,19 +24,7 @@ export const deletePost = async (postId: string) => {
 
 export const getPopularPosts = async () => {
   try {
-    const popularPosts = await db.post.findMany({
-      orderBy: {
-        comments: {
-          _count: 'desc',
-        },
-      },
-      take: 5,
-      include: {
-        user: true,
-        comments: true,
-        topic: true,
-      },
-    });
+    const popularPosts = await getPopularPostsFromDB();
     return popularPosts;
   } catch (error) {
     console.error('Error fetching popular posts:', error);
